Consolidation: clippy-clean, runtime Python parity, per-pillar docs
Three-part consolidation pass.

Clippy: `cargo clippy --all-targets` is now warning-clean (default +
python feature). Mechanical fixes (redundant casts, manual
is_multiple_of, map_or, redundant closures, doc-list indentation,
sort_by_key, while-let, ...) plus a small set of justified suppressions:
should_implement_trait on the partial-order cmp/eq (Surreal/Ordinal/
Rational/NumberGame/Game are deliberately not std-Ord/Eq), a crate-level
needless_range_loop allow for the index-parallel matrix code, and a
type_complexity allow on one test fixture.

Python bindings: brought to runtime-friendly parity with the Rust core.
Newly bound — games: the kernel outcome surface (outcomes/p_positions/
scoring_values, mex), the misère/octal surface (nim_canonical,
misere_nim_p_predicted, nim_moves, octal_moves, AbstractGame, Quotient,
octal_misere_quotient), and the loopy graph engine (LoopyGraph,
loopy_nim_values); forms: classify_real/classify_complex, WittClassG.mul,
hilbert_product, isotropy_over_adeles/AdelicIsotropy; scalars: Nimber
pow/__pow__/frobenius/sqrt. Const-generic backends (Qp/Zp/Fp/Laurent/...)
and closure-generic fns stay Rust-only by design — documented in
src/py/AGENTS.md, not silent gaps. demo.py covers the new surface.

Docs: split the monolithic AGENTS.md into per-pillar files (scalar,
clifford, forms, games, py, linalg), each with the full file-by-file
detail and layer-specific "things that look like bugs"; the root
AGENTS.md is now a linking summary. README rewritten around the
symmetries across the number systems (place vs characteristic, char-0 ↔
char-2 classifiers, surreal No ↔ ordinal On₂, the 2×2 functor table,
local ↔ global). Fixed doc drift (loopy.rs, symplectic.rs were missing).

+++ examples/tour.rs
@@ -91,7 +91,7 @@ fn main() {
 
     rule("char-0 classifier — Cl(p,q) as a matrix algebra (companion to Arf)");
     let cl = |qs: &[i128]| {
-        let q = qs.iter().map(|&x| Surreal::from_int(x as i128)).collect();
+        let q = qs.iter().map(|&x| Surreal::from_int(x)).collect();
         classify_surreal(&Metric::diagonal(q)).unwrap().display()
     };
     println!("  Cl(0,2) = {}   (the quaternions)", cl(&[-1, -1]));
